Created ActionHandler

+++ src/main/java/de/c0debase/bot/listener/message/MessageReceiveListener.java
@@ -99,44 +99,14 @@ public class MessageReceiveListener extends ListenerAdapter {
                 AIRequest request = new AIRequest(event.getMessage().getContentRaw().replace(event.getGuild().getSelfMember().getAsMention(), ""));
                 AIResponse response = CodebaseBot.getInstance().getAiDataService().request(request);
 
-                if(response.getResult().getAction().contains("weather")){
-                    if(!response.getResult().getParameters().containsKey("address")){
-                        event.getChannel().sendMessage("Leider kann ich nicht verstehen, was du von mir möchtest.").queue();
-                        return;
-                    }
-                    JsonObject jsonObject = new JsonParser().parse(response.getResult().getParameters().get("address").toString()).getAsJsonObject();
-                    if(jsonObject.get("city") == null){
-                        event.getChannel().sendMessage("Leider konnte ich das Wetter in diesem Ort nicht herausfinden").queue();
-                    }
-                    CurrentWeatherOneLocationQuery currentWeatherOneLocationQuery = QueryBuilderPicker.pick()
-                            .currentWeather()
-                            .oneLocation()
-                            .byCityName(jsonObject.get("city").getAsString())
-                            .type(Type.ACCURATE)
-                            .language(Language.GERMAN)
-                            .responseFormat(ResponseFormat.JSON)
-                            .unitFormat(UnitFormat.METRIC)
-                            .build();
-                    CurrentWeather currentWeather = CodebaseBot.getInstance().getDataWeatherClient().getCurrentWeather(currentWeatherOneLocationQuery);
-                    EmbedBuilder embedBuilder = new EmbedBuilder();
-                    embedBuilder.setColor(Color.ORANGE);
-                    embedBuilder.setTitle("Wetter in " + currentWeather.getCityName(), "https://openweathermap.org/city/" + currentWeather.getCityId());
-                    embedBuilder.addField("Temperatur",currentWeather.getMainParameters().getTemperature() + "°C", false);
-                    embedBuilder.addField("Luftdruck", String.format("%.1f hPa", currentWeather.getMainParameters().getPressure()), false);
-                    embedBuilder.addField("Windgeschwindigkeit", String.format("%.1f km/h", currentWeather.getWind().getSpeed()), false);
-                    embedBuilder.addField("Luftfeuchtigkeit",  String.format("%.1f", currentWeather.getMainParameters().getHumidity()) + "%\n", false);
-                    if(!currentWeather.getWeather().isEmpty()){
-                        embedBuilder.setFooter(currentWeather.getWeather().get(0).getDescription(), event.getGuild().getIconUrl());
-                    }
-                    event.getTextChannel().sendMessage(embedBuilder.build()).queue();
+                if(CodebaseBot.getInstance().getActionHandlerMap().containsKey(response.getResult().getAction().split("\\.")[0])){
+                    CodebaseBot.getInstance().getActionHandlerMap().get(response.getResult().getAction().split("\\.")[0]).handle(response, event.getMessage());
                     return;
                 }
-                if (response.getStatus().getCode() == 200) {
-                    if (response.getResult().getFulfillment().getSpeech().trim().isEmpty()) {
-                        event.getChannel().sendMessage("Leider kann ich nicht verstehen, was du von mir möchtest.").queue();
-                    } else {
-                        event.getChannel().sendMessage(response.getResult().getFulfillment().getSpeech().replace("@everyone", "@ everyone").replace("@here", "@ here")).queue();
-                    }
+                if (response.getStatus().getCode() == 200 && !response.getResult().getFulfillment().getSpeech().trim().isEmpty()) {
+                    event.getChannel().sendMessage(response.getResult().getFulfillment().getSpeech().replace("@everyone", "@ everyone").replace("@here", "@ here")).queue();
+                } else {
+                    event.getChannel().sendMessage("Leider kann ich nicht verstehen, was du von mir möchtest.").queue();
                 }
             } catch (AIServiceException e) {
                 e.printStackTrace();
@@ -146,7 +116,7 @@ public class MessageReceiveListener extends ListenerAdapter {
         if (event.getMessage().getTextChannel().getName().equalsIgnoreCase("bot") && event.getMessage().getContentRaw().startsWith("!")) {
             CodebaseBot.getInstance().getCommandManager().execute(event.getMessage());
         } else if(!event.getTextChannel().getName().equalsIgnoreCase("friend")){
-            if(lastMessage.containsKey(event.getMember()) && lastMessage.get(event.getMember()).equalsIgnoreCase(event.getMessage().getContentRaw())){
+            if(lastMessage.containsKey(event.getMember()) && lastMessage.get(event.getMember()).equalsIgnoreCase(event.getMessage().getContentRaw()) && event.getMessage().getAttachments().isEmpty()){
                 event.getMessage().delete().queue();
                 return;
             } else {
